App/requirements.txt
opencv-contrib-python==4.2.0.34
opencv-python==4.2.0.34
pafy==0.5.5
Pillow==8.2.0
youtube-dl==2021.5.16